requirements.txt
aiohttp==3.8.6
beautifulsoup4==4.12.2
python-dotenv==1.0.0
numpy==1.26.2
pandas
cognite-sdk
msal